requirements.txt
gradio
google-generativeai
yfinance
python-dotenv
faiss-cpu
PyPDF2
streamlit
serpapi
gemini-api
pandas
XlsxWriter
chromadb
sentence_transformers
pypdf
langchain
tqdm
finnhub-python
google-search-results